% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{build_table_location}
\alias{build_table_location}
\title{Build pretty table of factor significance and level counts}
\usage{
build_table_location(df_table, print_columns = c("MITS", "non-MITS+DSS-only"))
}
\arguments{
\item{df_table}{output from mits_selection_factor_tables()}

\item{print_columns}{specific columns from the nested tibbles in the table column to print. Defaults to "MITS" and "non-MITS+DSS-only". "DSS-only" and "non-MITS" are the other two columns that can be printed.}
}
\description{
Build pretty table of factor significance and level counts
}
