% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{champs_web_report}
\alias{champs_web_report}
\title{Create html outputs for a given mortality rate/fraction analysis}
\usage{
champs_web_report(obj, path = tempfile())
}
\arguments{
\item{obj}{an object that comes from \code{\link[=get_rates_and_fractions]{get_rates_and_fractions()}}}

\item{path}{path to where the output html will be stored}
}
\description{
Create html outputs for a given mortality rate/fraction analysis
}
