% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cond_factor_tables}
\alias{cond_factor_tables}
\title{Compute counts of MITS cases and non-cases for a given condition by factors}
\usage{
cond_factor_tables(x, sites, catchments, champs_group)
}
\arguments{
\item{x}{processed CHAMPS dataset}

\item{sites}{a vector of site names to include in the calculations}

\item{catchments}{a vector of catchments to include in the calculations}

\item{champs_group}{CHAMPS group specifying the condition}
}
\description{
Compute counts of MITS cases and non-cases for a given condition by factors
}
