% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_dataset_directory}
\alias{create_dataset_directory}
\title{Create a dataset directory with the shell of a config.yaml file}
\usage{
create_dataset_directory(path)
}
\arguments{
\item{path}{path where input data will be stored (must be a non-existing
path or an existing directory)}
}
\description{
Create a dataset directory with the shell of a config.yaml file
}
