% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_interval}
\alias{get_interval}
\title{Get a credible interval for a fraction}
\usage{
get_interval(frac, tot, limit = 90)
}
\arguments{
\item{frac}{the fraction}

\item{tot}{the denominator used to compute the fraction}

\item{limit}{the percent confidence limit (0-100)}
}
\description{
Get a credible interval for a fraction
}
