% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{get_rate_frac_data}
\alias{get_rate_frac_data}
\title{Get data necessary to compute adjusted mortality fractions and rates}
\usage{
get_rate_frac_data(
  x,
  site,
  catchments = NULL,
  condition = NULL,
  icd10_regex = NULL,
  maternal = FALSE,
  causal_chain = TRUE,
  adjust_vars = NULL,
  factor_groups = NULL
)
}
\arguments{
\item{x}{an object read in from \code{\link[=read_and_validate_data]{read_and_validate_data()}}}

\item{site}{a site name to get rate data for}

\item{catchments}{a vector of catchments to include in the calculations}

\item{condition}{a CHAMPS condition (see \code{\link[=valid_conditions]{valid_conditions()}})}

\item{icd10_regex}{an optional regular expression specifying
ICD10 codes that define a condition}

\item{maternal}{Are we searching for a maternal condition or ICD10
specification? Default is FALSE.}

\item{causal_chain}{if TRUE, the causal chain is searched, if
FALSE, the underlying cause is searched}

\item{adjust_vars}{a vector of variables to adjust by}

\item{factor_groups}{A named list that specifies how to group factors}
}
\description{
Get data necessary to compute adjusted mortality fractions and rates
}
\note{
One or both of \code{icd10_regex} and \code{condition} must be specified
}
