% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{get_rates_and_fractions}
\alias{get_rates_and_fractions}
\title{Get rate and fraction data}
\usage{
get_rates_and_fractions(
  x,
  sites = NULL,
  catchments = NULL,
  group_catchments = TRUE,
  condition = NULL,
  icd10_regex = NULL,
  cond_name = condition,
  causal_chain = TRUE,
  adjust_vars_override = NULL,
  factor_groups = NULL,
  pval_cutoff = 0.1,
  prop_cutoff = 20
)
}
\arguments{
\item{x}{Processed CHAMPS dataset.}

\item{sites}{A vector of site names to include in the calculations. If NULL,
all sites with data (corresponding to value of use_dss) will be used.}

\item{catchments}{A vector of catchments to include in the calculations.
If NULL, all catchments with data (corresponding to value of use_dss)
will be used.}

\item{group_catchments}{Should all catchments within a site be grouped
together?}

\item{condition}{CHAMPS group specifying the condition}

\item{icd10_regex}{An optional regular expression specifying ICD10 codes
that define a condition.}

\item{cond_name}{The name of the condition to use in outputs (e.g. if
the condition is "Congenital birth defects", cond_name could be "CBD").
Defaults to \code{condition} if not specified.}

\item{causal_chain}{Should the search for the condition be
across the causal chain?}

\item{adjust_vars_override}{An optional vector of adjustment
variables that will override the automatic adjustment. Cannot be
more than two variables and must include age.}

\item{factor_groups}{A named list that specifies how to group factors}

\item{pval_cutoff}{P-value for which variables are considered
for adjustment. Ignored if adjust_vars_override is specified.}

\item{use_dss}{Should the calculations be done only for catchments that}

\item{prop_na_cutoff}{Proportion of missingness for which
variables are considered for adjustment. Ignored if
adjust_vars_override is specified.}
}
\description{
Get rate and fraction data
}
\note{
One or both of \code{icd10_regex} and \code{condition} must be specified
}
