% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_site_info}
\alias{get_site_info}
\title{Get a list of sites and catchments with data start and end years}
\usage{
get_site_info(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Get a list of sites and catchments with data start and end years
}
