% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_condition.R
\name{has_champs_group_cc}
\alias{has_champs_group_cc}
\title{Check to see if a CHAMPS group is in the causal chain}
\usage{
has_champs_group_cc(ads, group)
}
\arguments{
\item{ads}{CHAMPS analysis dataset}

\item{group}{string representing a CHAMPS group}
}
\description{
Check to see if a CHAMPS group is in the causal chain
}
