% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_condition.R
\name{has_icd10}
\alias{has_icd10}
\title{Check to see if an ICD10 regular expression is in the causal chain}
\usage{
has_icd10(ads, rgx, cc = TRUE)
}
\arguments{
\item{ads}{CHAMPS analysis dataset}

\item{rgx}{regular expression representing ICD10 codes}

\item{cc}{if TRUE: searches across causal chain, if FALSE: searches in
underlying cause only}
}
\description{
Check to see if an ICD10 regular expression is in the causal chain
}
