% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_condition.R
\name{has_icd10_cc}
\alias{has_icd10_cc}
\title{Check to see if an ICD10 regular expression is in the causal chain}
\usage{
has_icd10_cc(ads, rgx)
}
\arguments{
\item{ads}{CHAMPS analysis dataset}

\item{rgx}{regular expression representing ICD10 codes}
}
\description{
Check to see if an ICD10 regular expression is in the causal chain
}
