% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_condition.R
\name{has_maternal_champs_group}
\alias{has_maternal_champs_group}
\title{Check to see if a CHAMPS group is in the main maternal cause}
\usage{
has_maternal_champs_group(ads, group)
}
\arguments{
\item{ads}{CHAMPS analysis dataset}

\item{group}{string representing a CHAMPS group}
}
\description{
Check to see if a CHAMPS group is in the main maternal cause
}
