% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_condition.R
\name{has_maternal_icd10}
\alias{has_maternal_icd10}
\title{Check to see if an ICD10 regular expression is in the main maternal cause}
\usage{
has_maternal_icd10(ads, rgx)
}
\arguments{
\item{ads}{CHAMPS analysis dataset}

\item{rgx}{regular expression representing ICD10 codes}
}
\description{
Check to see if an ICD10 regular expression is in the main maternal cause
}
