% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{make_results}
\alias{make_results}
\title{View the results of a TODO}
\usage{
make_results(obj, path = tempfile())
}
\arguments{
\item{obj}{TODO}

\item{path}{TODO}
}
\description{
View the results of a TODO
}
