% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_calc.R
\name{mits_factor_tables}
\alias{mits_factor_tables}
\title{Compute counts of MITS and non-MITS+DSS-only deaths by factors}
\usage{
mits_factor_tables(
  x,
  sites = NULL,
  catchments = NULL,
  factor_groups = NULL,
  ctch_info = NULL
)
}
\arguments{
\item{x}{Processed CHAMPS dataset.}

\item{sites}{A vector of site names to include in the calculations. If NULL,
all sites with data will be used.}

\item{catchments}{A vector of catchments to include in the calculations.
If NULL, all catchments with data will be used.}

\item{factor_groups}{A named list that specifies how to group factors}

\item{ctch_info}{An object that tracks information about catchments in the data.
This is used internally and should be ignored by users.}
}
\description{
Compute counts of MITS and non-MITS+DSS-only deaths by factors
}
