% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{process_data}
\alias{process_data}
\title{Process a set of CHAMPS data files}
\usage{
process_data(x, start_year, end_year)
}
\arguments{
\item{x}{an object read in from \code{\link[=read_and_validate_data]{read_and_validate_data()}}}

\item{start_year}{start year for filtering the data}

\item{end_year}{end year for filtering the data}
}
\description{
Process a set of CHAMPS data files
}
