% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{rates_and_fractions_wide}
\alias{rates_and_fractions_wide}
\title{Wrapper for get_rates_and_fractions()}
\usage{
rates_and_fractions_wide(inputs, dat)
}
\arguments{
\item{inputs}{is a list object with of the following named items
in the list - Catchment, Site, Age, Condition , UC_or_CC, DSS.
The UC_or_CC can be 'UC' or 'CC' and DS is 'Yes' or 'No'
get_results_wide}

\item{dat}{is the object returned from process_data()}
}
\description{
Uses the get_rates_and_fractions() function but outputs only the
rates and fractions in one row instead of multiple tables in a list.
}
