% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{rates_fracs_to_df}
\alias{rates_fracs_to_df}
\title{Wrapper for get_rates_and_fractions()}
\usage{
rates_fracs_to_df(dat)
}
\arguments{
\item{dat}{A list of outputs from either get_rates_and_fractions() or
\code{batch_rates_and_fractions()}.}
}
\description{
Converts detailed output \code{get_rates_and_fractions()} to a
data frame of the major output statistics.
}
