% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_and_validate_data}
\alias{read_and_validate_data}
\title{Read and validate data files from a specified directory}
\usage{
read_and_validate_data(data_path)
}
\arguments{
\item{data_path}{path containing config.yaml and data files}
}
\description{
Read and validate data files from a specified directory
}
