% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_adjust_decision.R
\name{table_adjust_decision}
\alias{table_adjust_decision}
\title{Selected variables decision table}
\usage{
table_adjust_decision(
  obj,
  alpha_value = 0.1,
  pmissing = 20,
  percent_digits = 1
)
}
\arguments{
\item{obj}{an object that comes from \code{\link[=get_rates_and_fractions]{get_rates_and_fractions()}}}

\item{alpha_value}{the threshold to compare against the p-value.
Cells are colored based on this value}

\item{pmissing}{the percent missing allowed for a factor to be considered
in the decision table. Cells are colored using this value.}

\item{percent_digits}{the number of digits to round the percent missing to.}
}
\description{
Selected variables decision table
}
