% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_overview.R
\name{table_overview}
\alias{table_overview}
\title{Create an overview table of results}
\usage{
table_overview(obj)
}
\arguments{
\item{obj}{an object \url{from get_rates_and_fractions}}
}
\description{
Create an overview table of results
}
