% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_maternal_conditions}
\alias{valid_maternal_conditions}
\title{Get a list of valid maternal conditions}
\usage{
valid_maternal_conditions(x)
}
\arguments{
\item{x}{an object read in from \code{\link[=read_and_validate_data]{read_and_validate_data()}}}
}
\description{
Get a list of valid maternal conditions
}
