% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_top10.R
\name{batch_rates_and_fractions}
\alias{batch_rates_and_fractions}
\title{Calculate rates and fraction of an input list of values.}
\usage{
batch_rates_and_fractions(inputs_csv, dat_folder, start_year, end_year)
}
\arguments{
\item{inputs_csv}{is a path to your file.}

\item{dat_folder}{is a path for use with read_and_validate_data()}

\item{start_year}{is the start_year argument passed to
read_and_validate_data()}

\item{end_year}{is the end_year argument passed to
read_and_validate_data()}
}
\description{
Allows user to provide an input table in CSV format
to itteratively calculate results using get_rates_and_fractions()
with the output returned in one row per set of inputs.
}
