% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_rates_and_fractions}
\alias{batch_rates_and_fractions}
\title{Calculate rates and fraction of an input list of values.}
\usage{
batch_rates_and_fractions(x, inputs_csv)
}
\arguments{
\item{x}{Processed CHAMPS dataset.}

\item{inputs_csv}{Path to a csv file specifying each scenario to run.}
}
\description{
Allows user to provide an input table in CSV format
to iteratively calculate results using get_rates_and_fractions()
with the output returned in one row per set of inputs.
}
