% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{combine_decision_tables}
\alias{combine_decision_tables}
\title{Combine mits_selection_factor_tables() and cond_factor_tables() output to one decision dataframe}
\usage{
combine_decision_tables(tables_dat)
}
\arguments{
\item{tables_dat}{a named list object with a table from mits_selection_factor_tables() and a table from cond_factor_tables().  The names assigned to each table will be used in naming the table columns.}
}
\description{
Combine mits_selection_factor_tables() and cond_factor_tables() output to one decision dataframe
}
