% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_calc.R
\name{cond_factor_tables}
\alias{cond_factor_tables}
\title{Compute counts of MITS cases and non-cases for a given condition by factors}
\usage{
cond_factor_tables(
  x,
  sites = NULL,
  catchments = NULL,
  factor_groups = NULL,
  condition = NULL,
  icd10_regex = NULL,
  cond_name_short = condition[1],
  causal_chain = TRUE
)
}
\arguments{
\item{x}{Processed CHAMPS dataset.}

\item{sites}{A vector of site names to include in the calculations. If NULL,
all sites with data will be used.}

\item{catchments}{A vector of catchments to include in the calculations.
If NULL, all catchments with data will be used.}

\item{factor_groups}{A named list that specifies how to group factors}

\item{condition}{CHAMPS group specifying the condition}

\item{icd10_regex}{An optional regular expression specifying ICD10 codes
that define a condition.}

\item{cond_name_short}{The name of the condition to use in outputs (e.g. if
the condition is "Congenital birth defects", cond_name_short could be "CBD").
Defaults to value of \code{condition} if not specified. Required if only
\code{icd10_regex} is specified.}

\item{causal_chain}{if TRUE, the causal chain is searched, if
FALSE, the underlying cause is searched}
}
\description{
Compute counts of MITS cases and non-cases for a given condition by factors
}
