% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{get_rates_and_fractions}
\alias{get_rates_and_fractions}
\title{Get rate and fraction data}
\usage{
get_rates_and_fractions(
  x,
  sites = NULL,
  catchments = NULL,
  condition = NULL,
  icd10_regex = NULL,
  maternal = FALSE,
  cond_name_short = condition[1],
  causal_chain = TRUE,
  adjust_vars_override = NULL,
  factor_groups = NULL,
  pval_cutoff = 0.1,
  pct_na_cutoff = 20
)
}
\arguments{
\item{x}{Processed CHAMPS dataset.}

\item{sites}{A vector of site names to include in the calculations. If NULL,
all sites with data will be used.}

\item{catchments}{A vector of catchments to include in the calculations.
If NULL, all catchments with data will be used.}

\item{condition}{CHAMPS group specifying the condition}

\item{icd10_regex}{An optional regular expression specifying ICD10 codes
that define a condition.}

\item{maternal}{Are we searching for a maternal condition or ICD10
specification? Default is FALSE.}

\item{cond_name_short}{The name of the condition to use in outputs (e.g. if
the condition is "Congenital birth defects", cond_name_short could be "CBD").
Defaults to \code{condition} if not specified.}

\item{causal_chain}{if TRUE, the causal chain is searched, if
FALSE, the underlying cause is searched}

\item{adjust_vars_override}{An optional vector of adjustment
variables that will override the automatic adjustment. Cannot be
more than two variables and must include age.}

\item{factor_groups}{A named list that specifies how to group factors}

\item{pval_cutoff}{P-value for which variables are considered
for adjustment. Ignored if adjust_vars_override is specified.}

\item{pct_na_cutoff}{Proportion of missingness for which
variables are considered for adjustment. Ignored if
adjust_vars_override is specified.}
}
\description{
Get rate and fraction data
}
\note{
One or both of \code{icd10_regex} and \code{condition} must be specified
}
