% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mits_selection_factor_tables}
\alias{mits_selection_factor_tables}
\title{Compute counts of MITS and non-MITS+DSS-only deaths by factors}
\usage{
mits_selection_factor_tables(x, sites, catchments)
}
\arguments{
\item{x}{processed CHAMPS dataset}

\item{sites}{a vector of site names to include in the calculations}

\item{catchments}{a vector of catchments to include in the calculations}
}
\description{
Compute counts of MITS and non-MITS+DSS-only deaths by factors
}
