% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rates_fracs.R
\name{plot_rates_fracs}
\alias{plot_rates_fracs}
\title{Create a plot of rates or fractions}
\usage{
plot_rates_fracs(
  obj,
  type = "frac",
  height = "100\%",
  width = 800,
  plotly = TRUE
)
}
\arguments{
\item{obj}{an object from \code{\link[=get_rates_and_fractions]{get_rates_and_fractions()}}}

\item{type}{one of "frac" or "rate"}

\item{height}{height of plot if using plotly}

\item{width}{width of plot if using plotly}

\item{plotly}{should the result be a plotly graph?}
}
\description{
Create a plot of rates or fractions
}
