% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rates_fracs.R
\name{plot_rates_fracs}
\alias{plot_rates_fracs}
\title{Create a plot of rates or fractions}
\usage{
plot_rates_fracs(obj, type = "frac", plotly = TRUE)
}
\arguments{
\item{obj}{object (TODO)}

\item{type}{one of "frac" or "rate"}

\item{plotly}{should the result be a plotly graph?}
}
\description{
Create a plot of rates or fractions
}
