% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_factor_sig.R
\name{table_factor_sig_stats}
\alias{table_factor_sig_stats}
\title{Build pretty table of factor significance and level counts}
\usage{
table_factor_sig_stats(
  obj,
  which = c("mits", "cond"),
  print_columns = c("MITS", "non-MITS+DSS-only"),
  percent_digits = 1
)
}
\arguments{
\item{obj}{output from \code{get_rates_and_fractions()} or
\code{batch_get_rates_and_fractions()} for one site.}

\item{which}{one of "mits" or "cond" to specify which adjustment statistics
to display}

\item{print_columns}{specific columns from the nested tibbles in the table
column to print. Defaults to "MITS" and "non-MITS+DSS-only". "DSS-only" and
"non-MITS" are the other two columns that can be printed.}

\item{percent_digits}{the number of digits to round the percent columns to.
Defaults to 1.}
}
\description{
Build pretty table of factor significance and level counts
}
